move DnsCache into its own module
